Moved SparkUtils to SparkVersionUtils, as well as compatibility methods.
