+++ connector/src/main/scala/com/vertica/spark/datasource/v2/VerticaDatasourceV2Read.scala
@@ -17,12 +17,10 @@ import com.typesafe.scalalogging.Logger
 import com.vertica.spark.config.{DistributedFilesystemReadConfig, LogProvider, ReadConfig}
 import com.vertica.spark.datasource.core.{DSConfigSetupInterface, DSReader, DSReaderInterface}
 import com.vertica.spark.datasource.json.{JsonBatchFactory, VerticaJsonScan}
-import com.vertica.spark.util.compatibilities.DSReadCompatibilityTools
 import com.vertica.spark.util.error.{ConnectorError, ConnectorException, ErrorHandling, InitialSetupPartitioningError}
 import com.vertica.spark.util.pushdown.PushdownUtils
 import com.vertica.spark.util.schema.ComplexTypesSchemaTools
-import com.vertica.spark.util.spark.SparkUtils
-import com.vertica.spark.util.version.Version
+import com.vertica.spark.util.version.{SparkVersionTools, Version}
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.connector.expressions.aggregate._
 import org.apache.spark.sql.connector.read._
@@ -68,7 +66,7 @@ class VerticaScanBuilder(config: ReadConfig, readConfigSetup: DSConfigSetupInter
 
   protected val ctTools: ComplexTypesSchemaTools = new ComplexTypesSchemaTools()
 
-  protected val compatibility = new DSReadCompatibilityTools
+  protected val sparkVersionTools = new SparkVersionTools()
 
 /**
   * Builds the class representing a scan of a Vertica table
@@ -180,9 +178,9 @@ class VerticaScanBuilderWithPushdown(config: ReadConfig, readConfigSetup: DSConf
   }
 
   private def getGroupByColumns(aggregation: Aggregation): Array[StructField] = {
-    val sparkVersion = SparkUtils.getVersion(SparkUtils.getVersionString).getOrElse(Version(3,3))
-    compatibility
-      .getGroupByExpressions(sparkVersion, aggregation)
+    val sparkVersion = sparkVersionTools.getVersion.getOrElse(Version(3,3))
+    sparkVersionTools
+      .getCompatibleGroupByExpressions(sparkVersion, aggregation)
       .map(expr => StructField(expr.describe, getColType(expr.describe), nullable = false, Metadata.empty))
   }
 }
