+++ connector/src/main/scala/com/vertica/spark/datasource/v2/VerticaDatasourceV2Table.scala
@@ -17,10 +17,8 @@ import cats.data.Validated.{Invalid, Valid}
 import com.vertica.spark.config.{LogProvider, ReadConfig}
 import com.vertica.spark.datasource.core.{DSConfigSetupInterface, DSReadConfigSetup, DSWriteConfigSetup}
 import com.vertica.spark.datasource.v2
-import com.vertica.spark.util.compatibilities.DSTableCompatibilityTools
 import com.vertica.spark.util.error.{ErrorHandling, ErrorList}
-import com.vertica.spark.util.spark.SparkUtils
-import com.vertica.spark.util.version.Version
+import com.vertica.spark.util.version.{SparkVersionTools, Version}
 import org.apache.spark.sql.connector.catalog.{SupportsRead, SupportsWrite, Table, TableCapability}
 import org.apache.spark.sql.connector.read.ScanBuilder
 import org.apache.spark.sql.connector.write.{LogicalWriteInfo, WriteBuilder}
@@ -40,7 +38,7 @@ class VerticaTable(caseInsensitiveStringMap: CaseInsensitiveStringMap, readSetup
   // Cache the scan builder so we don't build it twice
   var scanBuilder : Option[VerticaScanBuilder] = None
 
-  val compatibility = new DSTableCompatibilityTools
+  val sparkVersionTools = new SparkVersionTools
 
   /**
    * A name to differentiate this table from other tables
@@ -88,9 +86,8 @@ class VerticaTable(caseInsensitiveStringMap: CaseInsensitiveStringMap, readSetup
           case Valid(cfg) => cfg
         }
         logger.debug("Config loaded")
-        val sparkVersion = SparkUtils.getVersion(SparkUtils.getVersionString).getOrElse(Version(3,2))
-        val scanBuilder = compatibility
-          .makeVerticaScanBuilder(sparkVersion, config, readSetupInterface)
+        val sparkVersion = sparkVersionTools.getVersion.getOrElse(Version(3,2))
+        val scanBuilder = sparkVersionTools.makeCompatibleVerticaScanBuilder(sparkVersion, config, readSetupInterface)
         this.scanBuilder = Some(scanBuilder)
         scanBuilder
     }
