--- connector/src/main/scala/com/vertica/spark/util/compatibilities/DSReadCompatibilityTools.scala
@@ -1,35 +1,0 @@
-// (c) Copyright [2020-2021] Micro Focus or one of its affiliates.
-// Licensed under the Apache License, Version 2.0 (the "License");
-// You may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.vertica.spark.util.compatibilities
-
-import com.vertica.spark.util.reflections.ReflectionTools
-import com.vertica.spark.util.version.Version
-import org.apache.spark.sql.connector.expressions.aggregate.Aggregation
-import org.apache.spark.sql.connector.expressions.Expression
-
-/**
- * Contains utility methods to support backward compatibility
- * */
-class DSReadCompatibilityTools(reflection: ReflectionTools = new ReflectionTools) {
-
-  def getGroupByExpressions(sparkVersion: Version, aggObj: Aggregation): Array[Expression] = {
-    if(sparkVersion.lessThan(Version(3, 3))){
-      // $COVERAGE-OFF$
-      reflection.aggregationInvokeMethod[Array[Expression]](aggObj, "groupByColumns")
-      // $COVERAGE-ON$
-    } else {
-      aggObj.groupByExpressions()
-    }
-  }
-}
